rename launch_instance() to launch_install_instance()
There is a need to launch an instance that does not kick off an install. While the current method can technically do this, it still logs a new instance being spawned to do a network install. It was easiest to just rename this method and create a new one for the more general case.

Change-Id: I61d24cdc453eadb279f5a9d4b687b4247da5084c